PANDORA Robotics Team - GUI
This is the GUI developed for monitoring, manipulation and data
acquisition from PANDORA’s robotics vehicle at the Aristotle University
of Thessaloniki. It was designed and implemented as part of the team’s
project in order to participate in the Robocup - RoboRescue
Competitions, starting from 2011 in Istanbul. The repository for the
whole project is here:
https://github.com/pandora-auth-ros-pkg/pandora-auth-ros-pkg?source=c